Fix bug where replacing fuzzy item with search would cause incorrect fuzzy behavior
Previously if a user replaced a fuzzy slot by search that did not exist
(nor its layout) in the bank, the item would incorrectly appear as a
placeholder if a variation was present in the bank.

For example, if a player had only Bastion Potion (4) in the bank (no
other placeholders), but then replaced a fuzzy slot with Bastion Potion (3),
The layout would incorrectly include a placeholder Bastion Potion (3)
and then the actual Bastion Potion (4) item, instead of just the Bastion
Potion (4) item taking the layout slot of the placeholder Bastion Potion
(3).

The bug is fixed by correctly tracking the ID in the layout if it was
added to the layout when it previously did not exist during
recalculation of the layout.