Disable hotkeys during chat box inputs
Previously hotkeys could interfere with bank searches. Now hotkeys will
be disabled whenever any sort of chat box input (PMs, bank search, etc.)
will unregister hotkeys. After the input box is closed, the hotkeys will
be registered again. A config option has been added to allow hotkeys to
persist even if a chat box input is present.

This includes a small refactor to handling of the registration of hot
keys, since the bank being open/closed also determines if they are
registered or not.

Note that ChatBoxItemSearch does not handle the registration of hotkeys
correctly. The underlying reason is that spawning a ChatBoxItemSearch
does not trigger a change to the VarClientInt `MESLAYERMODE`. This would
need to be fixed upstream I think.